test(ui): split document-title shell tests to satisfy max-lines
